PM refactor: initial sketch
